add harnesses to kani files
- some harnesses could not be moved to a kani specific file due to
  visibility
  - see src/epoch/gregorian.rs
