+++ src/epoch/gregorian.rs
@@ -799,201 +799,6 @@ mod ut_gregorian {
 #[cfg(kani)]
 mod kani_harnesses {
     use super::*;
-    #[kani::proof]
-    fn kani_harness_Epoch_compute_gregorian() {
-        let duration: Duration = kani::any();
-        let time_scale: TimeScale = kani::any();
-        Epoch::compute_gregorian(duration, time_scale);
-    }
-
-    #[kani::proof]
-    fn kani_harness_to_gregorian_str() {
-        let time_scale: TimeScale = kani::any();
-        let callee: Epoch = kani::any();
-        callee.to_gregorian_str(time_scale);
-    }
-
-    #[kani::proof]
-    fn kani_harness_to_gregorian_utc() {
-        let callee: Epoch = kani::any();
-        callee.to_gregorian_utc();
-    }
-
-    #[kani::proof]
-    fn kani_harness_to_gregorian_tai() {
-        let callee: Epoch = kani::any();
-        callee.to_gregorian_tai();
-    }
-
-    #[kani::proof]
-    fn kani_harness_Epoch_maybe_from_gregorian_tai() {
-        let year: i32 = kani::any();
-        let month: u8 = kani::any();
-        let day: u8 = kani::any();
-        let hour: u8 = kani::any();
-        let minute: u8 = kani::any();
-        let second: u8 = kani::any();
-        let nanos: u32 = kani::any();
-        Epoch::maybe_from_gregorian_tai(year, month, day, hour, minute, second, nanos);
-    }
-
-    #[kani::proof]
-    fn kani_harness_Epoch_maybe_from_gregorian() {
-        let year: i32 = kani::any();
-        let month: u8 = kani::any();
-        let day: u8 = kani::any();
-        let hour: u8 = kani::any();
-        let minute: u8 = kani::any();
-        let second: u8 = kani::any();
-        let nanos: u32 = kani::any();
-        let time_scale: TimeScale = kani::any();
-        Epoch::maybe_from_gregorian(year, month, day, hour, minute, second, nanos, time_scale);
-    }
-
-    #[kani::proof]
-    fn kani_harness_Epoch_from_gregorian_tai() {
-        let year: i32 = kani::any();
-        let month: u8 = kani::any();
-        let day: u8 = kani::any();
-        let hour: u8 = kani::any();
-        let minute: u8 = kani::any();
-        let second: u8 = kani::any();
-        let nanos: u32 = kani::any();
-        Epoch::from_gregorian_tai(year, month, day, hour, minute, second, nanos);
-    }
-
-    #[kani::proof]
-    fn kani_harness_Epoch_from_gregorian_tai_at_midnight() {
-        let year: i32 = kani::any();
-        let month: u8 = kani::any();
-        let day: u8 = kani::any();
-        Epoch::from_gregorian_tai_at_midnight(year, month, day);
-    }
-
-    #[kani::proof]
-    fn kani_harness_Epoch_from_gregorian_tai_at_noon() {
-        let year: i32 = kani::any();
-        let month: u8 = kani::any();
-        let day: u8 = kani::any();
-        Epoch::from_gregorian_tai_at_noon(year, month, day);
-    }
-
-    #[kani::proof]
-    fn kani_harness_Epoch_from_gregorian_tai_hms() {
-        let year: i32 = kani::any();
-        let month: u8 = kani::any();
-        let day: u8 = kani::any();
-        let hour: u8 = kani::any();
-        let minute: u8 = kani::any();
-        let second: u8 = kani::any();
-        Epoch::from_gregorian_tai_hms(year, month, day, hour, minute, second);
-    }
-
-    #[kani::proof]
-    fn kani_harness_Epoch_maybe_from_gregorian_utc() {
-        let year: i32 = kani::any();
-        let month: u8 = kani::any();
-        let day: u8 = kani::any();
-        let hour: u8 = kani::any();
-        let minute: u8 = kani::any();
-        let second: u8 = kani::any();
-        let nanos: u32 = kani::any();
-        Epoch::maybe_from_gregorian_utc(year, month, day, hour, minute, second, nanos);
-    }
-
-    #[kani::proof]
-    fn kani_harness_Epoch_from_gregorian_utc() {
-        let year: i32 = kani::any();
-        let month: u8 = kani::any();
-        let day: u8 = kani::any();
-        let hour: u8 = kani::any();
-        let minute: u8 = kani::any();
-        let second: u8 = kani::any();
-        let nanos: u32 = kani::any();
-        Epoch::from_gregorian_utc(year, month, day, hour, minute, second, nanos);
-    }
-
-    #[kani::proof]
-    fn kani_harness_Epoch_from_gregorian_utc_at_midnight() {
-        let year: i32 = kani::any();
-        let month: u8 = kani::any();
-        let day: u8 = kani::any();
-        Epoch::from_gregorian_utc_at_midnight(year, month, day);
-    }
-
-    #[kani::proof]
-    fn kani_harness_Epoch_from_gregorian_utc_at_noon() {
-        let year: i32 = kani::any();
-        let month: u8 = kani::any();
-        let day: u8 = kani::any();
-        Epoch::from_gregorian_utc_at_noon(year, month, day);
-    }
-
-    #[kani::proof]
-    fn kani_harness_Epoch_from_gregorian_utc_hms() {
-        let year: i32 = kani::any();
-        let month: u8 = kani::any();
-        let day: u8 = kani::any();
-        let hour: u8 = kani::any();
-        let minute: u8 = kani::any();
-        let second: u8 = kani::any();
-        Epoch::from_gregorian_utc_hms(year, month, day, hour, minute, second);
-    }
-
-    #[kani::proof]
-    fn kani_harness_Epoch_from_gregorian() {
-        let year: i32 = kani::any();
-        let month: u8 = kani::any();
-        let day: u8 = kani::any();
-        let hour: u8 = kani::any();
-        let minute: u8 = kani::any();
-        let second: u8 = kani::any();
-        let nanos: u32 = kani::any();
-        let time_scale: TimeScale = kani::any();
-        Epoch::from_gregorian(year, month, day, hour, minute, second, nanos, time_scale);
-    }
-
-    #[kani::proof]
-    fn kani_harness_Epoch_from_gregorian_at_midnight() {
-        let year: i32 = kani::any();
-        let month: u8 = kani::any();
-        let day: u8 = kani::any();
-        let time_scale: TimeScale = kani::any();
-        Epoch::from_gregorian_at_midnight(year, month, day, time_scale);
-    }
-
-    #[kani::proof]
-    fn kani_harness_Epoch_from_gregorian_at_noon() {
-        let year: i32 = kani::any();
-        let month: u8 = kani::any();
-        let day: u8 = kani::any();
-        let time_scale: TimeScale = kani::any();
-        Epoch::from_gregorian_at_noon(year, month, day, time_scale);
-    }
-
-    #[kani::proof]
-    fn kani_harness_Epoch_from_gregorian_hms() {
-        let year: i32 = kani::any();
-        let month: u8 = kani::any();
-        let day: u8 = kani::any();
-        let hour: u8 = kani::any();
-        let minute: u8 = kani::any();
-        let second: u8 = kani::any();
-        let time_scale: TimeScale = kani::any();
-        Epoch::from_gregorian_hms(year, month, day, hour, minute, second, time_scale);
-    }
-
-    #[kani::proof]
-    fn kani_harness_is_gregorian_valid() {
-        let year: i32 = kani::any();
-        let month: u8 = kani::any();
-        let day: u8 = kani::any();
-        let hour: u8 = kani::any();
-        let minute: u8 = kani::any();
-        let second: u8 = kani::any();
-        let nanos: u32 = kani::any();
-        is_gregorian_valid(year, month, day, hour, minute, second, nanos);
-    }
 
     #[kani::proof]
     fn kani_harness_january_years() {
