+++ src/epoch/ops.rs
@@ -378,109 +378,3 @@ impl Hash for Epoch {
         self.time_scale.hash(state);
     }
 }
-
-#[cfg(kani)]
-mod kani_harnesses {
-    use super::*;
-    #[kani::proof]
-    fn kani_harness_min() {
-        let other: Epoch = kani::any();
-        let callee: Epoch = kani::any();
-        callee.min(other);
-    }
-
-    #[kani::proof]
-    fn kani_harness_max() {
-        let other: Epoch = kani::any();
-        let callee: Epoch = kani::any();
-        callee.max(other);
-    }
-
-    #[kani::proof]
-    fn kani_harness_floor() {
-        let duration: Duration = kani::any();
-        let callee: Epoch = kani::any();
-        callee.floor(duration);
-    }
-
-    #[kani::proof]
-    fn kani_harness_ceil() {
-        let duration: Duration = kani::any();
-        let callee: Epoch = kani::any();
-        callee.ceil(duration);
-    }
-
-    #[kani::proof]
-    fn kani_harness_round() {
-        let duration: Duration = kani::any();
-        let callee: Epoch = kani::any();
-        callee.round(duration);
-    }
-
-    #[kani::proof]
-    fn kani_harness_to_time_of_week() {
-        let callee: Epoch = kani::any();
-        callee.to_time_of_week();
-    }
-
-    #[kani::proof]
-    fn kani_harness_weekday_in_time_scale() {
-        let time_scale: TimeScale = kani::any();
-        let callee: Epoch = kani::any();
-        callee.weekday_in_time_scale(time_scale);
-    }
-
-    #[kani::proof]
-    fn kani_harness_weekday() {
-        let callee: Epoch = kani::any();
-        callee.weekday();
-    }
-
-    #[kani::proof]
-    fn kani_harness_weekday_utc() {
-        let callee: Epoch = kani::any();
-        callee.weekday_utc();
-    }
-
-    #[kani::proof]
-    fn kani_harness_next() {
-        let weekday: Weekday = kani::any();
-        let callee: Epoch = kani::any();
-        callee.next(weekday);
-    }
-
-    #[kani::proof]
-    fn kani_harness_next_weekday_at_midnight() {
-        let weekday: Weekday = kani::any();
-        let callee: Epoch = kani::any();
-        callee.next_weekday_at_midnight(weekday);
-    }
-
-    #[kani::proof]
-    fn kani_harness_next_weekday_at_noon() {
-        let weekday: Weekday = kani::any();
-        let callee: Epoch = kani::any();
-        callee.next_weekday_at_noon(weekday);
-    }
-
-    #[kani::proof]
-    fn kani_harness_previous() {
-        let weekday: Weekday = kani::any();
-        let callee: Epoch = kani::any();
-        callee.previous(weekday);
-    }
-
-    #[kani::proof]
-    fn kani_harness_previous_weekday_at_midnight() {
-        let weekday: Weekday = kani::any();
-        let callee: Epoch = kani::any();
-        callee.previous_weekday_at_midnight(weekday);
-    }
-
-    #[kani::proof]
-    fn kani_harness_previous_weekday_at_noon() {
-        let weekday: Weekday = kani::any();
-        let callee: Epoch = kani::any();
-        callee.previous_weekday_at_noon(weekday);
-    }
-}
